% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean data that contain punctuation}
\usage{
clean_data(data = top100)
}
\arguments{
\item{data}{the billboard dataset of all top songs from all time}
}
\value{
'clean_data' returns a dataframe with the removed punctuation
}
\description{
Clean data that contain punctuation
}
\examples{
library(billboard100)

# Clean any music dataset
clean_data(top100)


}
