% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_model.R
\name{music_model}
\alias{music_model}
\title{Build a regression of the music data}
\usage{
music_model(data = spotify, artists)
}
\arguments{
\item{data}{the spotify dataset of all songs}
}
\value{
'music_model' returns a linear regression model that predicts
}
\description{
Build a regression of the music data
}
\examples{
library(billboard100)

# Build a regression model of spotify data
music_model(data)


}
