% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{spotify_all}
\alias{spotify_all}
\title{Spotify data set}
\format{
\subsection{\code{spotify_all}}{

A data frame with 330,087 rows and 7 columns:
\describe{
\item{date}{Date of chart}
\item{rank}{Rank for the week}
\item{song}{Song name}
\item{artist}{Artist name}
\item{last_week}{Previous week ranking}
\item{peak_rank}{Highest rank reached of all time}
\item{weeks_on_board}{Number of weeks on the Hot 100 chart}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/dhruvildave/billboard-the-hot-100-songs}
}
\usage{
spotify_all
}
\description{
Track, album, and artist data along with Spotify music traits.
}
\keyword{datasets}
