% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{top100}
\alias{top100}
\title{Billboard Top 100 data set}
\format{
\subsection{\code{top100}}{

A data frame with 330,087 rows and 7 columns:
\describe{
\item{date}{Date of chart}
\item{rank}{Rank for the week}
\item{song}{Song name}
\item{artist}{Artist name}
\item{last_week}{Previous week ranking}
\item{peak_rank}{Highest rank reached of all time}
\item{weeks_on_board}{Number of weeks on the Hot 100 chart}
...
}
}
}
\source{
\url{https://www.kaggle.com/datasets/dhruvildave/billboard-the-hot-100-songs}
}
\usage{
top100
}
\description{
Top 100 Billboard songs from 08-04-1958 to 11-06-2021.
}
\keyword{datasets}
