% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n.R
\name{top_n}
\alias{top_n}
\title{Find top x songs of all time}
\usage{
top_n(data = top100, id, n)
}
\arguments{
\item{data}{music charts dataset}

\item{id}{variable to analyze}

\item{n}{number of top listings}
}
\value{
a dataframe with x number of rows of top songs over time
}
\description{
Find top x songs of all time
}
\examples{
library(billboard100)

# Find the top 10 songs of all time
top_n(data=top100, 10)

# Find the top 20 songs of all time
top_n(data = top100, 20)

}
