% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n_bb100.R
\name{top_n_bb100}
\alias{top_n_bb100}
\title{Find top x songs of all time}
\usage{
top_n_bb100(data = top100, x)
}
\arguments{
\item{data}{the billboard dataset of all top songs from all time}

\item{x}{the number that the user inputs to find the top number of songs}
}
\value{
'top_n_bb100' returns a dataframe with x number of rows of top songs over time
}
\description{
Find top x songs of all time
}
\examples{
library(billboard100)

# Find the top 10 songs of all time
top_n_bb100(data=top100, 10)

# Find the top 20 songs of all time
top_n_bb100(data = top100, 20)

}
