% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_function.R
\name{TopN_BB100}
\alias{TopN_BB100}
\title{Find top x songs of all time}
\usage{
TopN_BB100(data = BB100, x)
}
\arguments{
\item{data}{the billboard dataset of all top songs from all time used to calculate the top x number of songs}

\item{x}{the number that the user inputs to find the top number of songs}
}
\value{
'TopN_BB100' returns a dataframe with x number of rows of top songs over time based on the user input
}
\description{
Find top x songs of all time
}
\examples{
library(billboard_100)

# Find the top 10 songs of all time
TopN_BB100(data=BB100, 10)

# Find the top 20 songs of all time
TopN_BB100(data = BB100, 20)

}
