% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize}
\usage{
visualize(data)
}
\arguments{
\item{data}{dataset with two variables of interest, output of the top_n function}
}
\value{
ggplot geom_col object
}
\description{
Visualize
}
\examples{
example
library(singR)

#Visualize top 4 songs between 1990-01-01 and 1995-01-01
ex <- top_n(id = "song", n = 4, date_start = "1990-01-01", date_end = "1995-01-01")
visualize(ex)


}
