% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/music_model.R
\name{explicit_model}
\alias{explicit_model}
\title{Build a regression of the music data}
\usage{
explicit_model(data = spotify_all, i1)
}
\arguments{
\item{data}{the spotify dataset of all songs}
}
\value{
'explicit_model' returns a linear regression model that predicts
}
\description{
Build a regression of the music data
}
\examples{
library(billboard100)

# Build a regression model of spotify data
explicit_model(data)


}
