% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize}
\alias{visualize}
\title{Visualize}
\usage{
visualize(data)
}
\arguments{
\item{data}{dataset with two variables of interest, output of the top_n function}
}
\value{
ggplot object depicting the dataset
}
\description{
Visualize
}
